% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capture_sites_within}
\alias{capture_sites_within}
\title{Intersecting sites within Basin}
\usage{
capture_sites_within(x, tog)
}
\arguments{
\item{x}{List of POLYGONS}

\item{tog}{Data that was joined earlier.}
}
\value{
A POLYGON with other sites that intersect it
}
\description{
Intersecting sites within Basin
}
