% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_cleaning}
\alias{date_cleaning}
\title{Clean dates}
\usage{
date_cleaning(data)
}
\arguments{
\item{data}{A previously created joined POU and POD object.}
}
\value{
A sf object
}
\description{
Clean dates
}
