% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fs_logic}
\alias{fs_logic}
\title{FS Logic}
\usage{
fs_logic(data, admin_int)
}
\arguments{
\item{data}{A previously created joined POU and POD object.}

\item{admin_int}{A previously created administration intersected sf object.}
}
\value{
A sf object with fs logic of intersection
}
\description{
FS Logic
}
