% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\name{get_adminboundaries}
\alias{get_adminboundaries}
\title{Get USDA-Forest Service Administration Boundaries}
\usage{
get_adminboundaries(filter_geom, ...)
}
\arguments{
\item{filter_geom}{an object of class bbox, sfc or sfg used to filter query results based on a predicate function.}

\item{...}{Arguments to pass to \link\link{arcgislayers} package \code{arc_select} function.}
}
\description{
An area depicted as surface ownership parcels dissolved on the same ownership classification. Go to this URL for full metadata description: https://data.fs.usda.gov/geodata/edw/edw_resources/meta/S_USA.BasicOwnership.xml
}
