% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\name{get_consumptive_estimate}
\alias{get_consumptive_estimate}
\title{Get Consumptive Estimate}
\usage{
get_consumptive_estimate(data, local_path = NULL, return_everything = TRUE)
}
\arguments{
\item{data}{A sf POLYGON or MULTIPOLYGON object.}

\item{local_path}{A file path character string to the MT gdb for water rights POU and POD layers. 'Z:/some_path/wr_rights.gdb'}

\item{return_everything}{A logical for return spatial objects and results (TRUE) or just tabular results (FALSE).}
}
\description{
Get Consumptive Estimate
}
