% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\name{get_flowmet}
\alias{get_flowmet}
\title{Get FlowMet Data}
\usage{
get_flowmet(filter_geom, ...)
}
\arguments{
\item{filter_geom}{an object of class bbox, sfc or sfg used to filter query results based on a predicate function.}

\item{...}{Arguments to pass to \link\link{arcgislayers} package \code{arc_select} function.}
}
\description{
A map service depicting modeled streamflow metrics from the historical time period (1977-2006) in the United States.
In addition to standard NHD attributes, the streamflow datasets include metrics on mean daily flow (annual and seasonal),
flood levels associated with 1.5-year, 10-year, and 25-year floods; annual and decadal minimum weekly flows and date of
minimum weekly flow, center of flow mass date; baseflow index, and average number of winter floods. These files and
additional information are available on the project website, https://www.fs.usda.gov/rm/boise/AWAE/projects/modeled_stream_flow_metrics.shtml. Streams without flow metrics (null values) were removed from this dataset to improve display speed; to see all stream lines, use an NHD flowline dataset.
}
