% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\name{get_mtwr}
\alias{get_mtwr}
\title{Get Montana Water Rights Data}
\usage{
get_mtwr(filter_geom, layer, local_path = NULL)
}
\arguments{
\item{filter_geom}{an object of class bbox be used for clipping.}

\item{layer}{Layer name within the \verb{https://ftpgeoinfo.msl.mt.gov/Data/Spatial/NonMSDI/DNRC_WR/MTWaterRights.gdb.zip} gdb.}

\item{local_path}{A file path character string to the MT gdb for water rights POU and POD layers. 'Z:/some_path/wr_rights.gdb'}
}
\description{
Get Montana Water Rights Data
}
