% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pod_basins}
\alias{get_pod_basins}
\title{Get POD basins}
\usage{
get_pod_basins(data, crs)
}
\arguments{
\item{data}{A previously filter flowmet object}

\item{crs}{A \code{sf::st_crs()} object.}
}
\description{
Get POD basins
}
