% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_consumptive}
\alias{plot_consumptive}
\title{Plot Consumptive Outputs}
\usage{
plot_consumptive(data)
}
\arguments{
\item{data}{A previously created \code{get_consumptive_estimate()} list.}
}
\description{
Plot Consumptive Outputs
}
